**Bug fixed
I had to understand how the script in package.json, 'npm run build' transpiles my ts files to js and copy them to '\www' folder and then creates a zip file of the www folder when our server will read from the zip file as specified in the config.yml file generated by eb create.

The error i got was "cannot find module '../../src/aws'" The server could not find the module because of there is no src folder existing in the zip file. So i now had to change the path to ./../aws for the server to be able to get the module. it did not complain about the other modules that had path because the folder path actually existed in the server.

This was a good learning experience

I now understand why i should have
  "main": "server.js",
"start": "node server.js",
"prod": "tsc && node ./www/server.js", // if this line work perfectly in your local server, you are sure it will be perfect when you deploy to aws and set the environment variables in env configuration as well

To deploy
- Test your app in 'npm run prod' first
eb init
- npm run build
- eb create

To deploy changes
- npm run build
- eb deploy